requirements.txt
astroid==2.3.3
certifi==2019.11.28
cffi==1.14.0
colorama==0.4.3
cryptography==2.8
decorator==4.4.2
dnspython==1.16.0
future==0.18.2
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pycparser==2.20
pylint==2.4.4
pymongo==3.10.1
PySocks==1.7.1
python-telegram-bot==12.4.2
six==1.14.0
tornado==6.0.4
wrapt==1.11.2
